+++ core/caching/dict_cache.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import dataclasses
 import datetime
 from collections.abc import MutableMapping
-from threading import Lock
 
 from core.caching.cache import Cache
 from core.util import date_util
@@ -18,15 +17,13 @@ class DictCache(Cache):
     def __init__(self, isPrivate: bool = False) -> None:
         super().__init__(isPrivate=isPrivate)
         self._entries: MutableMapping[str, DictCache.CacheEntry] = {}
-        self._lock = Lock()
 
     async def set(self, key: str, value: str, expirySeconds: float) -> bool:
-        with self._lock:
-            self._entries[key] = DictCache.CacheEntry(
-                value=value,
-                expiryDate=date_util.datetime_from_now(seconds=expirySeconds),
-            )
-            return True
+        self._entries[key] = DictCache.CacheEntry(
+            value=value,
+            expiryDate=date_util.datetime_from_now(seconds=expirySeconds),
+        )
+        return True
 
     def _internal_get(self, key: str) -> str | None:
         entry = self._entries.get(key)
@@ -38,15 +35,13 @@ class DictCache(Cache):
         return entry.value
 
     async def get(self, key: str) -> str | None:
-        with self._lock:
-            return self._internal_get(key=key)
+        return self._internal_get(key=key)
 
     async def delete(self, key: str) -> bool:
-        with self._lock:
-            ret = self._internal_get(key=key)
-            if ret:
-                del self._entries[key]
-            return ret is not None
+        ret = self._internal_get(key=key)
+        if ret:
+            del self._entries[key]
+        return ret is not None
 
     def can_store_complex_objects(self) -> bool:  # pylint: disable=no-self-use
         return True